Postman Testing for API

+++ redis-demo/src/main/java/com/learn/redis/redisdemo/dao/ProgrammerRepository.java
@@ -26,7 +26,7 @@ public interface ProgrammerRepository {
     void saveHash(Programmer programmer);
     void updateHash(Programmer programmer);
     Map<Long,Programmer> findAllHash();
-    Programmer findInHash(int id);
-    void deleteHash(int id);
+    Programmer findInHash(Long id);
+    void deleteHash(Long id);
 
 }
